requirements.txt
flask
openai
requests
beautifulsoup4
python-dotenv
apscheduler
google-search-results